+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2015-10-19
+
+Update to 2.1 :
+
+- Fix #10, #11 : Give the ability to call a php function in the yaml file,
+  eg : `%function_name:arg1,arg2%`
+
 # 2015-08-06
 
 Update to 2.0 :